pyproject.toml
[project]
name = "questarctica"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=0.114.0",
    "langchain-openai>=0.3.14",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.9.1",
    "pygithub>=2.6.1",
    "pymongo>=4.12.0",
    "python-dotenv>=1.1.0",
    "todoist-api-python>=3.0.1",
]
